separate objects, add interfaces
