+++ helper/ldaputil/client_test.go
